+++ examples/simple/override_simple_theme/fobi_themes.py
@@ -11,7 +11,7 @@ class MySimpleTheme(SimpleTheme):
     html_classes = ['my-simple-theme',]
     base_view_template = 'override_simple_theme/base_view.html'
     form_ajax = 'override_simple_theme/snippets/form_ajax.html'
-    #form_snippet_template_name = 'override_simple_theme/snippets/form_snippet.html'
+    form_snippet_template_name = 'override_simple_theme/snippets/form_snippet.html'
 
 
 # It's important to set the `force` argument to True, in