+++ scripts/reinstall_django_1_5.sh
@@ -1,0 +1,3 @@
+reset
+./scripts/uninstall.sh
+./scripts/install_django_1_5.sh