+++ src/fobi/__init__.py
@@ -1,6 +1,6 @@
 __title__ = 'django-fobi'
-__version__ = '0.4.29'
-__build__ = 0x000032
+__version__ = '0.4.30'
+__build__ = 0x000033
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = '2014-2015 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'