+++ src/fobi/contrib/plugins/form_handlers/mail/fobi_form_handlers.py
@@ -1,22 +1,28 @@
+from __future__ import absolute_import
+
 __title__ = 'fobi.contrib.plugins.form_handlers.mail.fobi_form_handlers'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = 'Copyright (c) 2014 Artur Barseghyan'
+__copyright__ = 'Copyright (c) 2014-2015 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('MailHandlerPlugin',)
+
+from mimetypes import guess_type
+import os
 
 from six import string_types
 
 from django.utils.translation import ugettext_lazy as _
-from django.core.mail import send_mail
+#from django.core.mail import send_mail
 from django.template.loader import render_to_string
 from django.conf import settings
 
 from fobi.base import (
     FormHandlerPlugin, form_handler_plugin_registry, get_processed_form_data
 )
-from fobi.helpers import safe_text
+from fobi.helpers import safe_text, extract_file_path
 from fobi.contrib.plugins.form_handlers.mail import UID
 from fobi.contrib.plugins.form_handlers.mail.forms import MailForm
+from fobi.contrib.plugins.form_handlers.mail.helpers import send_mail
 
 class MailHandlerPlugin(FormHandlerPlugin):
     """
@@ -59,13 +65,51 @@ class MailHandlerPlugin(FormHandlerPlugin):
                 safe_text(label), safe_text(cleaned_data[key]))
                 )
 
+        files = self._prepare_files(request, form)
+
         send_mail(
             safe_text(self.data.subject),
             "{0}\n\n{1}".format(safe_text(self.data.body), ''.join(rendered_data)),
             self.data.from_email,
             [self.data.to_email],
-            fail_silently = True
+            fail_silently = False,
+            attachments = files.values()
             )
+
+    def _prepare_files(self, request, form):
+        """
+        Prepares the files for being attached to the mail message.
+        """
+        files = {}
+
+        def process_path(file_path, imf):
+            """
+            Processes the file path and the file.
+            """
+            if file_path:
+                #if file_path.startswith(settings.MEDIA_URL):
+                #    file_path = file_path[1:]
+                #file_path = settings.PROJECT_DIR('../{0}'.format(file_path))
+                file_path = file_path.replace(
+                    settings.MEDIA_URL,
+                    os.path.join(settings.MEDIA_ROOT, '')
+                    )
+                mime_type = guess_type(imf.name)
+                files[field_name] = (
+                    imf.name,
+                    ''.join([c for c in imf.chunks()]),
+                    mime_type[0] if mime_type else ''
+                )
+
+        for field_name, imf in request.FILES.items():
+            try:
+                file_path = form.cleaned_data.get(field_name, '')
+                process_path(file_path, imf)
+            except Exception as e:
+                file_path = extract_file_path(imf.name)
+                process_path(file_path, imf)
+
+        return files
 
     def plugin_data_repr(self):
         """