+++ src/fobi/contrib/plugins/form_handlers/mail/helpers.py
@@ -1,0 +1,34 @@
+from __future__ import absolute_import
+
+__title__ = 'fobi.contrib.plugins.form_handlers.mail.helpers'
+__author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
+__copyright__ = 'Copyright (c) 2014-2015 Artur Barseghyan'
+__license__ = 'GPL 2.0/LGPL 2.1'
+__all__ = ('send_mail',)
+
+from django.core.mail import get_connection
+from django.core.mail.message import EmailMultiAlternatives
+
+def send_mail(subject, message, from_email, recipient_list,
+              fail_silently=False, auth_user=None, auth_password=None,
+              connection=None, html_message=None, attachments=None):
+    """
+    Easy wrapper for sending a single message to a recipient list. All members
+    of the recipient list will see the other recipients in the 'To' field.
+
+    If auth_user is None, the EMAIL_HOST_USER setting is used.
+    If auth_password is None, the EMAIL_HOST_PASSWORD setting is used.
+
+    Note: The API for this method is frozen. New code wanting to extend the
+    functionality should use the EmailMessage class directly.
+    """
+    connection = connection or get_connection(username=auth_user,
+                                              password=auth_password,
+                                              fail_silently=fail_silently)
+    mail = EmailMultiAlternatives(subject, message, from_email, recipient_list,
+                                  connection=connection,
+                                  attachments=attachments)
+    if html_message:
+        mail.attach_alternative(html_message, 'text/html')
+
+    return mail.send()